# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Setmark Home Server",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/workouts": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Create Workout",
        "description": "Create a workout, or upsert-in-place if the caller supplied an `id` that\nalready exists for this user.\n\nbug-041: prior to this change, a POST with an `id` that matched an existing\nrow returned 500 (PK constraint violation). Seed scripts and any client\nthat replays a POST with stable IDs hit that path. Single-user-cutover\nphilosophy says \"same id = upsert, not crash\" \u2014 so we route an existing-id\nPOST through the same tree-replace logic PUT uses. New-id POSTs take the\nplain insert path.",
        "operationId": "create_workout_api_workouts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "List Workouts",
        "operationId": "list_workouts_api_workouts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "scheduled_date >= this (YYYY-MM-DD).",
              "title": "After"
            },
            "description": "scheduled_date >= this (YYYY-MM-DD)."
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Matches when tags_json contains this tag.",
              "title": "Tag"
            },
            "description": "Matches when tags_json contains this tag."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max rows to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max rows to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Rows to skip for pagination.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Rows to skip for pagination."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutRead"
                  },
                  "title": "Response List Workouts Api Workouts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{workout_id}": {
      "put": {
        "tags": [
          "workouts"
        ],
        "summary": "Update Workout",
        "operationId": "update_workout_api_workouts__workout_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "Get Workout",
        "operationId": "get_workout_api_workouts__workout_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityIntentIn": {
        "properties": {
          "activity_domain": {
            "type": "string",
            "enum": [
              "running",
              "cycling",
              "rowing",
              "swimming",
              "walking",
              "hiking",
              "functional_strength",
              "traditional_strength",
              "hiit",
              "mobility",
              "mixed_modal",
              "carry",
              "other"
            ],
            "title": "Activity Domain"
          },
          "environment": {
            "type": "string",
            "enum": [
              "indoor",
              "outdoor",
              "unspecified"
            ],
            "title": "Environment",
            "default": "unspecified"
          },
          "preservation_policy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "preserve_primary_activity",
                  "preserve_structure",
                  "preserve_elapsed_time",
                  "preserve_distance",
                  "preserve_mixed_modality"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preservation Policy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activity_domain"
        ],
        "title": "ActivityIntentIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrimitiveBlockIn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "repeat": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Repeat",
            "default": 1
          },
          "work_target": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveWorkTargetIn"
            },
            "type": "array",
            "title": "Work Target"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveSetIn"
            },
            "type": "array",
            "title": "Sets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sets"
        ],
        "title": "PrimitiveBlockIn"
      },
...
      "PrimitiveSetIn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "timing": {
            "$ref": "#/components/schemas/PrimitiveTimingIn"
          },
          "traversal": {
            "type": "string",
            "enum": [
              "sequential",
              "round_robin",
              "amrap"
            ],
            "title": "Traversal",
            "default": "sequential"
          },
          "repeat": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Repeat",
            "default": 1
          },
          "work_target": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveWorkTargetIn"
            },
            "type": "array",
            "title": "Work Target"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveSlotIn"
            },
            "type": "array",
            "title": "Slots"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "timing"
        ],
        "title": "PrimitiveSetIn"
      },
...
      "PrimitiveSlotIn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "work_target": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveWorkTargetIn"
            },
            "type": "array",
            "title": "Work Target"
          },
          "load": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrimitiveLoadIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "stimuli": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveStimulusIn"
            },
            "type": "array",
            "title": "Stimuli"
          },
          "post_rest_sec": {
            "type": "integer",
            "title": "Post Rest Sec",
            "default": 0
          },
          "is_warmup": {
            "type": "boolean",
            "title": "Is Warmup",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "exercise_id"
        ],
        "title": "PrimitiveSlotIn"
      },
      "PrimitiveStimulusIn": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "rir",
              "hr_zone"
            ],
            "title": "Type"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "PrimitiveStimulusIn"
      },
      "PrimitiveTimingIn": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "set_bounded",
              "time_bounded",
              "cap_bounded",
              "target_bounded"
            ],
            "title": "Mode"
          },
          "interval_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Sec"
          },
          "rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds"
          },
          "cap_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap Sec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "PrimitiveTimingIn"
      },
      "PrimitiveWorkTargetIn": {
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "reps",
              "duration",
              "distance",
              "rounds",
              "completion",
              "load_carried"
            ],
            "title": "Metric"
          },
          "value_form": {
            "type": "string",
            "enum": [
              "single",
              "range",
              "open"
            ],
            "title": "Value Form"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "role": {
            "type": "string",
            "enum": [
              "completion",
              "observation"
            ],
            "title": "Role"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metric",
          "value_form",
          "role"
        ],
        "title": "PrimitiveWorkTargetIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkoutCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "active",
              "completed",
              "skipped"
            ],
            "title": "Status",
            "default": "planned"
          },
          "source": {
            "type": "string",
            "enum": [
              "claude",
              "manual"
            ],
            "title": "Source",
            "default": "claude"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "activity_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityIntentIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "primitive_blocks": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveBlockIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Primitive Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "primitive_blocks"
        ],
        "title": "WorkoutCreate",
        "description": "Accepts a full primitive workout tree (blocks \u2192 sets \u2192 slots) in one request.\n\nuser_id is resolved from the bearer token, not the body (ADR-2026-04-17)."
      },
      "WorkoutRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "activity_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityIntentIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "primitive_blocks": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveBlockIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Primitive Blocks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "scheduled_date",
          "status",
          "source",
          "notes",
          "tags_json",
          "created_at",
          "updated_at",
          "completed_at",
          "primitive_blocks"
        ],
        "title": "WorkoutRead"
      },
...
      "WorkoutUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "planned",
                  "active",
                  "completed",
                  "skipped"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "activity_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityIntentIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "primitive_blocks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PrimitiveBlockIn"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Primitive Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkoutUpdate",
        "description": "Partial update. Blocks, if present, replace the full tree (simpler than diff)."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}